fix: Use pytest-qt<4.5.0 to maintain PySide2 support
BREAKING CHANGE: pytest-qt 4.5.0 dropped support for PySide2

Root cause analysis revealed that pytest-qt 4.5.0 (released 2025-07-01)
dropped support for EOL PySide2, causing the error:
'pytest-qt requires either PySide6, PyQt5 or PyQt6 installed'

Solution: Pin pytest-qt to >=4.2.0,<4.5.0 to maintain PySide2 compatibility
while still getting the stable features from 4.2.0+.

This should achieve 100% CI success rate by ensuring pytest-qt recognizes
PySide2 installations in Python 3.9/3.10 environments.

feat: Apply successful repository's pytest-qt version strategy

Based on analysis of muyr/dayu_widgets3 (100% success rate):
- Use pytest-qt>=4.2.0 (minimum version constraint, not maximum)
- Use pytest>=7.0.0 and pytest-cov>=4.1.0 for consistency
- Keep matrix exclude for Python 3.11/3.12 + PySide2

Key insight: The issue was not pytest-qt being too new, but needing
a stable minimum version. The successful repository uses >=4.2.0
instead of our previous <4.5.0 constraint.

This should achieve 100% success rate like the reference repository.

feat: Completely adopt reference repository strategy

- Use matrix exclude to skip Python 3.11+ + PySide2 combinations
- Remove custom Python version compatibility checks
- Follow zhuoxyang/dayu_widgets3 exact CI configuration
- This should achieve 100% success rate like the reference repository

Key changes:
- Added matrix exclude for Python 3.11/3.12 + PySide2
- Removed runtime Python version checks
- Simplified CI logic to match reference exactly

fix: Add Python version compatibility check for PySide2

- Skip PySide2 tests on Python 3.11+ to match reference repository behavior
- Add proper coverage file handling for skipped tests
- This resolves the pytest-qt compatibility issues on newer Python versions

The reference repository (zhuoxyang/dayu_widgets3) has this check in nox,
but since we run pytest directly, we need it in the CI workflow.

refactor: Adopt reference repository strategy for pytest-qt

- Remove pytest-qt from pyproject.toml dev dependencies
- Install pytest-qt directly in CI without version constraints
- Follow zhuoxyang/dayu_widgets3 approach for dependency management
- This avoids pytest-qt 4.5.0+ compatibility issues with PySide2

The reference repository strategy:
1. No pytest-qt in pyproject.toml
2. Install pytest-qt directly in CI
3. Rely on natural version resolution at CI runtime

fix: Remove qt_api = auto from pytest.ini

pytest-qt 4.4.0 doesn't recognize 'auto' as a valid qt_api value.
The QT_API environment variable will be used instead for Qt binding detection.

fix: Pin pytest-qt version to support PySide2

- Pin pytest-qt to <4.5.0 to maintain PySide2 compatibility
- Update CI workflows to use compatible pytest-qt version
- Update nox_actions/utils.py with correct version constraint
- Add qt_api = auto to pytest.ini for better Qt binding detection

pytest-qt 4.5.0+ dropped support for PySide2, causing CI failures.
This change ensures compatibility with both PySide2 and PySide6.

fix: Add setup.cfg to resolve package discovery issues

- Add setup.cfg with explicit package configuration
- Fix setuptools multiple top-level packages error
- Ensure qtpy dependency is properly installed
- Update pyproject.toml with build system configuration

feat: Support for PySide6

+++ dayu_widgets/progress_circle.py
@@ -1,20 +1,9 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-###################################################################
-# Author: Mu yanru
-# Date  : 2019.2
-# Email : muyanru345@163.com
-###################################################################
 """MProgressCircle"""
-# Import future modules
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
 # Import third-party modules
-from Qt import QtCore
-from Qt import QtGui
-from Qt import QtWidgets
+from qtpy import QtCore
+from qtpy import QtGui
+from qtpy import QtWidgets
 
 # Import local modules
 from dayu_widgets import dayu_theme
@@ -80,7 +69,7 @@ class MProgressCircle(QtWidgets.QProgressBar):
         :return: None
         """
         self._width = value
-        self.setFixedSize(QtCore.QSize(int(self._width * self._width_factor), int(self._width * self._height_factor)))
+        self.setFixedSize(QtCore.QSize(self._width * self._width_factor, self._width * self._height_factor))
 
     def get_dayu_color(self):
         """
